pyproject.toml
[project]
name = "cdn-upload-cli"
version = "0.1.0"
description = "CLI tool for uploading images to Cloudflare R2 CDN with automatic optimization"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "CDN Upload CLI"}
]
keywords = ["cdn", "upload", "cloudflare", "r2", "webp", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities",
]

dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pillow>=10.0.0",
    "boto3>=1.34.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "ffmpeg-python>=0.2.0",
    "pyperclip>=1.8.2",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
cdn-upload = "cdn_upload.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/cdn-upload-cli"
Repository = "https://github.com/yourusername/cdn-upload-cli"
Documentation = "https://github.com/yourusername/cdn-upload-cli#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cdn_upload"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]
